scripts/requirements.txt
azure-cosmos==4.5.1
azure-search-documents==11.4.0
azure-core==1.29.5
openai==1.3.0
python-dotenv==1.0.0
